+++ packages/espressocash_app/lib/features/onboarding/src/widgets/onboarding_notice.dart
@@ -1,3 +1,4 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -8,8 +9,8 @@ import '../../../../ui/button.dart';
 import '../../../accounts/models/account.dart';
 import '../../../accounts/models/ec_wallet.dart';
 import '../../../accounts/services/accounts_bloc.dart';
-import '../../onboarding.dart';
 import '../data/onboarding_repository.dart';
+import '../screens/onboarding_flow_screen.dart';
 
 class OnboardingNotice extends StatefulWidget {
   const OnboardingNotice({super.key});
@@ -94,4 +95,9 @@ class _Content extends StatelessWidget {
           const SizedBox(width: 24),
         ],
       );
+}
+
+extension OnboardingExt on BuildContext {
+  Future<bool?> launchOnboardingFlow() =>
+      router.push<bool?>(OnboardingFlowScreen.route());
 }
